fixed some minor bugs
it used to be that a user could be a member of the server, but !userpoints would return that they didn't exist. This has been fixed, as the command will now check if the user's id is present in the list of members of the server, as well as the point dictionary. If they are found, it states that they have 0 points.